Allow for /user-attachments in readme images
